add configs and fix code
