+++ projects/swiftkv/configs/llama-3.1-swiftkv-70b-instruct.yaml
@@ -1,22 +1,20 @@
 type: swiftkv
-code: train_qwen2.py
+code: ../train.py
 micro_batch_size: 1
 epochs: 1
 gradient_accumulation_steps: 1
 temperature: 2.0
-decoder_loss_mult: 0.0
 model:
-  name_or_path: Qwen/Qwen2.5-32B-Instruct
-  num_key_value_layers: 36
+  name_or_path: meta-llama/Llama-3.1-70B-Instruct
+  num_key_value_layers: 40
   key_value_group_size: 1
-  attn_implementation: flash_attention_2
 deepspeed:
   zero_optimization:
     stage: 3
 data:
   sources:
     - HuggingFaceH4/ultrachat_200k
-    - teknium/OpenHermes-2.5
+    - lmsys/lmsys-chat-1m
     - meta-math/MetaMathQA
     - ise-uiuc/Magicoder-OSS-Instruct-75K
   use_data_cache: true
@@ -25,17 +23,17 @@ data:
   num_proc: 16
   max_length: 8192
 logger:
-  level: INFO
-  #output_dir: "./"
-  #file_output_ranks: [0]
+  level: WARNING
+  output_dir: "./"
+  file_output_ranks: [0]
 scheduler:
   warmup_ratio: 0.05
 optimizer:
   betas: [0.9,0.999]
   weight_decay: 0.0
-  lr: 0.0001
+  lr: 4e-4
 checkpoint:
   - type: huggingface
     save_every_n_epochs: 1
-    output_dir: /checkpoint/aqiao/icml25/swiftkv-qwen2.5-32b-36l-1g-1e-4
+    output_dir: /data-fast/llama-3.1-swiftkv-70b-instruct
     save_end_of_training: true
