--- projects/swiftkv/swiftkv-llama-70b.yaml
@@ -1,36 +1,0 @@
-type: swiftkv
-micro_batch_size: 1
-epochs: 1
-gradient_accumulation_steps: 1
-temperature: 2.0
-decoder_loss_mult: 0.0
-model:
-  name_or_path: meta-llama/Llama-3.3-70B-Instruct
-  num_key_value_layers: 40
-  key_value_group_size: 1
-deepspeed:
-  zero_optimization:
-    stage: 3
-data:
-  sources:
-    - HuggingFaceH4/ultrachat_200k
-  use_data_cache: true
-  cache_processed_data: true
-  cache_dir: /data/data-cache
-  num_proc: 16
-  max_length: 8192
-logger:
-  level: WARNING
-  output_dir: "./"
-  file_output_ranks: [0]
-scheduler:
-  warmup_ratio: 0.05
-optimizer:
-  betas: [0.9,0.999]
-  weight_decay: 0.0
-  lr: 4e-4
-checkpoint:
-  - type: huggingface
-    save_every_n_epochs: 1
-    output_dir: /data/swiftkv-llama-70b
-    save_end_of_training: true
