--- projects/swiftkv/swiftkv-llama-8b.yaml
@@ -1,41 +1,0 @@
-type: swiftkv
-code: train_llama.py
-micro_batch_size: 1
-epochs: 1
-gradient_accumulation_steps: 1
-temperature: 2.0
-decoder_loss_mult: 0.0
-model:
-  name_or_path: meta-llama/Llama-3.1-8B-Instruct
-  num_key_value_layers: 16
-  key_value_group_size: 1
-  attn_implementation: flash_attention_2
-deepspeed:
-  zero_optimization:
-    stage: 2
-data:
-  sources:
-    - HuggingFaceH4/ultrachat_200k
-    - teknium/OpenHermes-2.5
-    - meta-math/MetaMathQA
-    - ise-uiuc/Magicoder-OSS-Instruct-75K
-  use_data_cache: true
-  cache_processed_data: true
-  cache_dir: /data-fast/data-cache
-  num_proc: 16
-  max_length: 8192
-logger:
-  level: INFO
-  #output_dir: "./"
-  #file_output_ranks: [0]
-scheduler:
-  warmup_ratio: 0.05
-optimizer:
-  betas: [0.9,0.999]
-  weight_decay: 0.0
-  lr: 0.0002
-checkpoint:
-  - type: huggingface
-    save_every_n_epochs: 1
-    output_dir: /data-fast/swiftkv-llama-8b-fa2
-    save_end_of_training: true
