--- projects/swiftkv/swiftkv-qwen2.5-32b-48l-2g.yaml
@@ -1,41 +1,0 @@
-type: swiftkv
-code: train_qwen2.py
-micro_batch_size: 1
-epochs: 1
-gradient_accumulation_steps: 1
-temperature: 2.0
-decoder_loss_mult: 0.0
-model:
-  name_or_path: Qwen/Qwen2.5-32B-Instruct
-  num_key_value_layers: 48
-  key_value_group_size: 2
-  attn_implementation: flash_attention_2
-deepspeed:
-  zero_optimization:
-    stage: 3
-data:
-  sources:
-    - HuggingFaceH4/ultrachat_200k
-    - teknium/OpenHermes-2.5
-    - meta-math/MetaMathQA
-    - ise-uiuc/Magicoder-OSS-Instruct-75K
-  use_data_cache: true
-  cache_processed_data: true
-  cache_dir: /data-fast/data-cache
-  num_proc: 16
-  max_length: 8192
-logger:
-  level: INFO
-  #output_dir: "./"
-  #file_output_ranks: [0]
-scheduler:
-  warmup_ratio: 0.05
-optimizer:
-  betas: [0.9,0.999]
-  weight_decay: 0.0
-  lr: 0.0002
-checkpoint:
-  - type: huggingface
-    save_every_n_epochs: 1
-    output_dir: /checkpoint/aqiao/icml25/swiftkv-qwen2.5-32b-48l-2g
-    save_end_of_training: true
